Add test tags to every test.
The reason for this is that if we want to use TEMPLATE_TEST_CASE(), then a test
tag is necessary; and so at that point just for the sake of consistency, it
makes sense to add test tags to everything.

There is no functionality change in this commit---it's entirely housekeeping.