Add detection for FP16 and BF16 types.
If we are using C++23, then we can depend on stdfloat types and our job is easy.
If we are not, we can still detect native FP16 and BF16 support for gcc and
clang in some situations.

If the hardware supports it, and we have C++23, then these types are enabled by
default.  If we don't have C++23, the user can specify ARMA_FORCE_USE_FP16 and
ARMA_FORCE_USE_BF16 to try and force detection.  Warnings will be printed if (1)
the support for bf16 and fp16 is found but detected to be software emulated
(slow!) or (2) no suitable bf16 or fp16 type is found, in which case support is
disabled.

This commit also adds typedefs for relevant matrix types, as well as a new
handful of traits and restrictors that will be useful in enabling support
throughout.  Note that FP16 and BF16 are *not* BLAS types!