Add or disable BF16/FP16 support for specific operations.
In some cases, we have to restrict the operations further as they depend on BLAS
types; in other cases, we can use Armadillo's non-BLAS emulated support to
implement the operation.

In a few places, it is necessary to manually cast types, as __fp16(float) (and
similar) will always cause compilation warnings with gcc and clang, in case the
user did not intend for it.  But we intend it.