Add templated tests for BF16 and FP16 types.
We define the list of types we want to test in utils.hpp, and then use
TEMPLATE_TEST_CASE() throughout to test various different Armadillo operations.
This does not test every single possible operation in Armadillo---so support may
not be complete as of this commit---but this tests basically everything that's
already in the test suite that would be applicable to BF16 and FP16 types.